+++ main_MLP_Dis_posEnc.py
@@ -7,8 +7,8 @@ import sys
 import numpy as np
 import pandas as pd
 from misc import MLPDisentanglePos, SupervisedContrastiveLoss, normalizeFeatures, loadData, cumulate_EMA, evaluation
-from misc import sim_dist_specifc_loss_spc, sup_contra_Cplus2_classes
-from sklearn.metrics import f1_score, confusion_matrix, accuracy_score, cohen_kappa_score
+from misc import sim_dist_specifc_loss_spc, sup_contra_Cplus2_classes, plot_confusion_matrix
+from sklearn.metrics import f1_score, confusion_matrix, accuracy_score, cohen_kappa_score, ConfusionMatrixDisplay
 import torch.nn.functional as F
 #import warnings
 #warnings.filterwarnings('ignore')
@@ -31,7 +31,8 @@ normalize_features = True
 use_valid = False # use validation dataset
 momentum_ema = .95
 data_path = '../LU22_final_shared/'
-model_name = "model_MLP_Dis_posEnc_" + suffix + '_Lev' + str(pred_level) + '_' + str(epochs) + 'ep_seed' + str(rng_seed) + '.pth'
+config_details = "MLP_Dis_posEnc_" + suffix + '_Lev' + str(pred_level) + '_' + str(epochs) + 'ep_seed' + str(rng_seed)
+model_name = "model_" + config_details + '.pth'
 
 print(f'(Random seed set to {rng_seed})')
 torch.manual_seed(rng_seed)
@@ -98,8 +99,6 @@ test_dataloader = DataLoader(test_dataset, shuffle=False, batch_size=1024)
 
 ######## Model training
 device = 'cuda' if torch.cuda.is_available() else 'cpu'
-print(f'Training model on {device}...')
-
 # Model
 # 1) Random Forest (theirs)
 # model = RandomForestClassifier(n_estimators=130, criterion='gini', max_depth= None, min_samples_leaf=2, max_features= 'sqrt' , oob_score=True)
@@ -109,115 +108,120 @@ print(f'Training model on {device}...')
 model = MLPDisentanglePos(n_classes).to(device)
 # model = TempCNNDisentangleV4(n_classes).to(device)
 
+if os.path.isfile(model_name):
+    print(f'Loading model weights (previously trained)...')
+    model.load_state_dict(torch.load(model_name,map_location=torch.device(device)))
+else:
+    print(f'Training model on {device}...')
 
-learning_rate = 0.0001
-loss_fn = nn.CrossEntropyLoss()
-scl = SupervisedContrastiveLoss()
+    learning_rate = 0.0001
+    loss_fn = nn.CrossEntropyLoss()
+    scl = SupervisedContrastiveLoss()
 
-optimizer = torch.optim.AdamW(params=model.parameters(), lr=learning_rate)
+    optimizer = torch.optim.AdamW(params=model.parameters(), lr=learning_rate)
 
-ema_weights = None
+    ema_weights = None
 
-for epoch in range(epochs):
-    start = time.time()
-    model.train()
-    tot_loss = 0.0
-    domain_loss = 0.0
-    contra_tot_loss = 0.0
-    den = 0
+    for epoch in range(epochs):
+        start = time.time()
+        model.train()
+        tot_loss = 0.0
+        domain_loss = 0.0
+        contra_tot_loss = 0.0
+        den = 0
 
-    for x_batch, y_batch, dom_batch, coord_batch in train_dataloader:
-        if x_batch.shape[0] != training_batch_size:
-            continue
+        for x_batch, y_batch, dom_batch, coord_batch in train_dataloader:
+            if x_batch.shape[0] != training_batch_size:
+                continue
 
-        x_batch = x_batch.to(device)
-        y_batch = y_batch.to(device)
-        dom_batch = dom_batch.to(device)
-        coord_batch = coord_batch.to(device)
-        optimizer.zero_grad()
-        pred, inv_emb, spec_emb_d, spec_d_pred, inv_emb_n1, spec_emb_n1, inv_fc_feat, spec_fc_feat = model(x_batch, coord_batch)
+            x_batch = x_batch.to(device)
+            y_batch = y_batch.to(device)
+            dom_batch = dom_batch.to(device)
+            coord_batch = coord_batch.to(device)
+            optimizer.zero_grad()
+            pred, inv_emb, spec_emb_d, spec_d_pred, inv_emb_n1, spec_emb_n1, inv_fc_feat, spec_fc_feat = model(x_batch, coord_batch)
 
-        ohe_label = F.one_hot(y_batch,num_classes=n_classes).cpu().detach().numpy()
-        ohe_dom = F.one_hot(dom_batch,num_classes=2).cpu().detach().numpy()
+            ohe_label = F.one_hot(y_batch,num_classes=n_classes).cpu().detach().numpy()
+            ohe_dom = F.one_hot(dom_batch,num_classes=2).cpu().detach().numpy()
 
-        ##### DOMAIN CLASSIFICATION #####
-        loss_ce_spec_dom = loss_fn(spec_d_pred, dom_batch)
+            ##### DOMAIN CLASSIFICATION #####
+            loss_ce_spec_dom = loss_fn(spec_d_pred, dom_batch)
 
-        ##### MIXED MAINFOLD & CONTRASTIVE LEARNING ####
-        
-        cl_labels_npy = y_batch.cpu().detach().numpy()
-        #dummy_labels_npy = np.ones_like(cl_labels_npy) * n_classes
-        #y_mix_labels = np.concatenate([ cl_labels_npy , dummy_labels_npy , cl_labels_npy],axis=0)
-        y_mix_labels = np.concatenate([ cl_labels_npy , cl_labels_npy],axis=0)
-        #y_mix_labels = np.concatenate([ cl_labels_npy , dummy_labels_npy],axis=0)
-        
-        
-        #DOMAIN LABEL FOR DOMAIN-CLASS SPECIFIC EMBEDDING and DOMAIN SPECIFIC EMBEDDING IS 0 OR 1 
-        spec_dc_dom_labels = dom_batch.cpu().detach().numpy()
-        #DOMAIN LABEL FOR INV EMBEDDING IS 2
-        inv_dom_labels = np.ones_like(spec_dc_dom_labels) * 2
-
-        dom_mix_labels = np.concatenate([inv_dom_labels, spec_dc_dom_labels],axis=0)
-        
-        joint_embedding = torch.concat([inv_emb, spec_emb_d])
-        mixdl_loss_supContraLoss = sim_dist_specifc_loss_spc(joint_embedding, y_mix_labels, dom_mix_labels, scl, epoch)
-        
-        joint_embedding_n1 = torch.concat([inv_emb_n1, spec_emb_n1])
-        mixdl_loss_supContraLoss_n1 = sim_dist_specifc_loss_spc(joint_embedding_n1, y_mix_labels, dom_mix_labels, scl, epoch)
-
-        joint_embedding_fc_feat = torch.concat([inv_fc_feat, spec_fc_feat])
-        mixdl_loss_supContraLoss_fc = sim_dist_specifc_loss_spc(joint_embedding_fc_feat, y_mix_labels, dom_mix_labels, scl, epoch)
-        # mixdl_loss_supContraLoss_fc = sup_contra_Cplus2_classes(joint_embedding_fc_feat, y_mix_labels, dom_mix_labels, scl, epoch)
-        
-        contra_loss = mixdl_loss_supContraLoss_fc #+ mixdl_loss_supContraLoss_n1 + mixdl_loss_supContraLoss 
-
-        ####################################
-
-        loss = loss_fn(pred, y_batch) + contra_loss + loss_ce_spec_dom
-
-        loss.backward() # backward pass: backpropagate the prediction loss
-        optimizer.step() # gradient descent: adjust the parameters by the gradients collected in the backward pass
-        tot_loss+= loss.cpu().detach().numpy()
-        contra_tot_loss+= contra_loss.cpu().detach().numpy()
-        den+=1.
-
-    end = time.time()
-
-    # Evaluation
-    with torch.no_grad():
-        if use_valid:
-            pred_valid, labels_valid = evaluation(model, valid_dataloader, device)
-            f1_val = f1_score(labels_valid, pred_valid, average="weighted")
+            ##### MIXED MAINFOLD & CONTRASTIVE LEARNING ####
             
-            eval_test = (f1_val > valid_f1)
-            if f1_val > valid_f1:
+            cl_labels_npy = y_batch.cpu().detach().numpy()
+            #dummy_labels_npy = np.ones_like(cl_labels_npy) * n_classes
+            #y_mix_labels = np.concatenate([ cl_labels_npy , dummy_labels_npy , cl_labels_npy],axis=0)
+            y_mix_labels = np.concatenate([ cl_labels_npy , cl_labels_npy],axis=0)
+            #y_mix_labels = np.concatenate([ cl_labels_npy , dummy_labels_npy],axis=0)
+            
+            
+            #DOMAIN LABEL FOR DOMAIN-CLASS SPECIFIC EMBEDDING and DOMAIN SPECIFIC EMBEDDING IS 0 OR 1 
+            spec_dc_dom_labels = dom_batch.cpu().detach().numpy()
+            #DOMAIN LABEL FOR INV EMBEDDING IS 2
+            inv_dom_labels = np.ones_like(spec_dc_dom_labels) * 2
+
+            dom_mix_labels = np.concatenate([inv_dom_labels, spec_dc_dom_labels],axis=0)
+            
+            joint_embedding = torch.concat([inv_emb, spec_emb_d])
+            mixdl_loss_supContraLoss = sim_dist_specifc_loss_spc(joint_embedding, y_mix_labels, dom_mix_labels, scl, epoch)
+            
+            joint_embedding_n1 = torch.concat([inv_emb_n1, spec_emb_n1])
+            mixdl_loss_supContraLoss_n1 = sim_dist_specifc_loss_spc(joint_embedding_n1, y_mix_labels, dom_mix_labels, scl, epoch)
+
+            joint_embedding_fc_feat = torch.concat([inv_fc_feat, spec_fc_feat])
+            mixdl_loss_supContraLoss_fc = sim_dist_specifc_loss_spc(joint_embedding_fc_feat, y_mix_labels, dom_mix_labels, scl, epoch)
+            # mixdl_loss_supContraLoss_fc = sup_contra_Cplus2_classes(joint_embedding_fc_feat, y_mix_labels, dom_mix_labels, scl, epoch)
+            
+            contra_loss = mixdl_loss_supContraLoss_fc #+ mixdl_loss_supContraLoss_n1 + mixdl_loss_supContraLoss 
+
+            ####################################
+
+            loss = loss_fn(pred, y_batch) + contra_loss + loss_ce_spec_dom
+
+            loss.backward() # backward pass: backpropagate the prediction loss
+            optimizer.step() # gradient descent: adjust the parameters by the gradients collected in the backward pass
+            tot_loss+= loss.cpu().detach().numpy()
+            contra_tot_loss+= contra_loss.cpu().detach().numpy()
+            den+=1.
+
+        end = time.time()
+
+        # Evaluation
+        with torch.no_grad():
+            if use_valid:
+                pred_valid, labels_valid = evaluation(model, valid_dataloader, device)
+                f1_val = f1_score(labels_valid, pred_valid, average="weighted")
+                
+                eval_test = (f1_val > valid_f1)
+                if f1_val > valid_f1:
+                    torch.save(model.state_dict(), model_name)
+                    valid_f1 = f1_val
+            else:
                 torch.save(model.state_dict(), model_name)
-                valid_f1 = f1_val
-        else:
-            torch.save(model.state_dict(), model_name)
-            eval_test = (epoch%10 == 0)
+                eval_test = (epoch%10 == 0)
 
 
-        ####################### EMA #####################################
-        if epoch >= epochs/2:        
-            ema_weights = cumulate_EMA(model, ema_weights, momentum_ema)
-            # current_state_dict = model.state_dict()        
-            # model.load_state_dict(ema_weights)
-            # pred_test, labels_test = evaluation(model, test_dataloader, device)
-            # f1_ema = f1_score(labels_test, pred_test, average="weighted")
-            # model.load_state_dict(current_state_dict)
-        ####################### EMA #####################################
+            ####################### EMA #####################################
+            if epoch >= epochs/2:        
+                ema_weights = cumulate_EMA(model, ema_weights, momentum_ema)
+                # current_state_dict = model.state_dict()        
+                # model.load_state_dict(ema_weights)
+                # pred_test, labels_test = evaluation(model, test_dataloader, device)
+                # f1_ema = f1_score(labels_test, pred_test, average="weighted")
+                # model.load_state_dict(current_state_dict)
+            ####################### EMA #####################################
 
-        if eval_test:
-            pred_test, labels_test = evaluation(model, test_dataloader, device)
-            # acc = accuracy_score(labels_test, pred_test)
-            f1 = f1_score(labels_test, pred_test, average="weighted")
-            print("Epoch %d (%.2fs): train loss %.4f, contrastive %.4f. F1 on TEST %.2f"%(epoch, (end-start), tot_loss/den, contra_tot_loss/den, 100*f1))
-            #print(confusion_matrix(labels_test, pred_test))        
-        else:
-            print("Epoch %d (%.2fs): train loss %.4f, contrastive %.4f"%(epoch, (end-start), tot_loss/den, contra_tot_loss/den))
-    
-        sys.stdout.flush()
+            if eval_test:
+                pred_test, labels_test = evaluation(model, test_dataloader, device)
+                # acc = accuracy_score(labels_test, pred_test)
+                f1 = f1_score(labels_test, pred_test, average="weighted")
+                print("Epoch %d (%.2fs): train loss %.4f, contrastive %.4f. F1 on TEST %.2f"%(epoch, (end-start), tot_loss/den, contra_tot_loss/den, 100*f1))
+                #print(confusion_matrix(labels_test, pred_test))        
+            else:
+                print("Epoch %d (%.2fs): train loss %.4f, contrastive %.4f"%(epoch, (end-start), tot_loss/den, contra_tot_loss/den))
+        
+            sys.stdout.flush()
 
 
 
@@ -232,11 +236,28 @@ print(f'>>> Scenario: Data {suffix}, Level {pred_level}')
 print(f'MLP+dis TEST perf: Acc={acc*100:.2f}, F1={f1*100:.2f}')
 np.set_printoptions(precision=2)
 print(f'F1 per-class: {f1_perclass*100}')
-# print(confusion_matrix(labels_test, pred_test))
+
+# Confusion matrix
+conf_matrix = confusion_matrix(labels_test, pred_test)
+# save to csv
+cm_filename = "confusion_matrix_" + config_details
+df_conf_matrix = pd.DataFrame(conf_matrix)
+df_conf_matrix.to_csv(cm_filename + '.csv', index=False)
+# Plot
+text = True if pred_level==1 else False # Display values on each cell in confusion matrix
+cm_title = f'Confusion Matrix ({suffix}, level {pred_level})'
+plot_confusion_matrix(conf_matrix, title=cm_title, filename = cm_filename, text=text)
+# Plot normalized per-row
+conf_matrix = confusion_matrix(labels_test, pred_test, normalize='true')
+plot_confusion_matrix(conf_matrix, title=cm_title, filename = cm_filename + '_norm', normalized=True, text=text)
 
 ### Final assessment - EMA model
-model.load_state_dict(ema_weights)
-torch.save(model.state_dict(), 'EMA' + model_name)
+if os.path.isfile('EMA' + model_name):
+    print(f'Loading model weights (previously trained)...')
+    model.load_state_dict(torch.load('EMA' + model_name,map_location=torch.device(device)))
+else:
+    model.load_state_dict(ema_weights)
+    torch.save(model.state_dict(), 'EMA' + model_name)
 
 pred_test, labels_test = evaluation(model, test_dataloader, device)
 acc = accuracy_score(labels_test, pred_test)
