+++ main_RF.py
@@ -3,8 +3,8 @@ import sys
 import numpy as np
 import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
-from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score, f1_score, cohen_kappa_score, classification_report 
-from misc import normalizeFeatures, loadData
+from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score, f1_score, cohen_kappa_score, ConfusionMatrixDisplay
+from misc import normalizeFeatures, loadData, plot_confusion_matrix
 import joblib
 
 
@@ -16,8 +16,9 @@ rng_seed = int(sys.argv[2]) if len(sys.argv) > 2 else 42
 # gapfill: all samples but cloud free features only
 suffix = sys.argv[3] if len(sys.argv) > 3 else 'prime' #'prime' or 'gapfill'
 data_path = '../LU22_final_shared/'
-model_name = "model_RF_" + suffix + '_Lev' + str(pred_level) + '_seed' + str(rng_seed) + '.pth'
-normalize_features = True
+config_details = "RF_" + suffix + '_Lev' + str(pred_level) + '_seed' + str(rng_seed)"
+model_name = "model_" + config_details + '.pth'
+normalize_features = False
 
 print(f'(Random seed set to {rng_seed})')
 # torch.manual_seed(rng_seed)
@@ -40,19 +41,24 @@ print(f'test_data shape: {test_data.shape}')
 print(f'test_label shape: {test_label.shape}')
 
 ######## Model training
-print('Training model...')
-if suffix == 'prime':
-    n_estimators=130 if pred_level==1 else 150
-else: # 'gapfill'
-    n_estimators=110 if pred_level==1 else 115
+if os.path.isfile(model_name):
+    print(f'Loading model weights (previously trained)...')
+    model = joblib.load(model_name)
+else:
+    print('Training model...')
+    if suffix == 'prime':
+        n_estimators=130 if pred_level==1 else 150
+    else: # 'gapfill'
+        n_estimators=110 if pred_level==1 else 115
 
-model = RandomForestClassifier(n_estimators=130, criterion='gini', max_depth= None, min_samples_leaf=2, max_features= 'sqrt' , oob_score=True)
-model.fit(train_data, train_label)
+    model = RandomForestClassifier(n_estimators=130, criterion='gini', max_depth= None, min_samples_leaf=2, max_features= 'sqrt' , oob_score=True)
+    model.fit(train_data, train_label)
+
+    # save
+    joblib.dump(model, model_name)
 
 y_pred = model.predict(test_data)
 
-# save
-joblib.dump(model, model_name)
 
 ### Final assessment
 acc = accuracy_score(test_label, y_pred)
@@ -64,4 +70,17 @@ print(f'>>> Scenario: Data {suffix}, Level {pred_level}')
 print(f'RF TEST perf: Acc={acc*100:.2f}, F1={f1*100:.2f}')
 np.set_printoptions(precision=2)
 print(f'F1 per-class: {f1_perclass*100}')
-# print(confusion_matrix(y_pred, test_label))
+
+# Confusion matrix
+conf_matrix = confusion_matrix(test_label, y_pred)
+# save to csv
+cm_filename = "confusion_matrix_" + config_details
+df_conf_matrix = pd.DataFrame(conf_matrix)
+df_conf_matrix.to_csv(cm_filename + '.csv', index=False)
+# Plot
+text = True if pred_level==1 else False # Display values on each cell in confusion matrix
+cm_title = f'Confusion Matrix ({suffix}, level {pred_level})'
+plot_confusion_matrix(conf_matrix, title=cm_title, filename = cm_filename, text=text)
+# Plot normalized per-row
+conf_matrix = confusion_matrix(test_label, y_pred, normalize='true')
+plot_confusion_matrix(conf_matrix, title=cm_title, filename = cm_filename + '_norm', normalized=True, text=text)
